chore: rename CreateRecipeVIew file and fix nested navigation container issue
