+++ final/HelloRecipe/HelloRecipe/Views/RecipeListView.swift
@@ -11,44 +11,42 @@ struct RecipeListView: View {
     @EnvironmentObject var recipeViewModel: RecipeViewModel
 
     var body: some View {
-        NavigationView {
-            ScrollView {
-                HStack {
-                    Spacer()
-                    NavigationLink(destination: CreateReceipt()) {
-                        Image(systemName: "plus")
-                            .imageScale(.large)
-                            .foregroundStyle(.tint)
-                        Text("Add New")
-                    }.padding(.trailing)
-                }
-
-                VStack(spacing: 24) {
-                    if !recipeViewModel.cookingRecipes.isEmpty {
-                        RecipeSection(
-                            title: RecipeStatus.cooking.displayTitle,
-                            recipes: recipeViewModel.cookingRecipes
-                        )
-                    }
-
-                    if !recipeViewModel.toCookRecipes.isEmpty {
-                        RecipeSection(
-                            title: RecipeStatus.toCook.displayTitle,
-                            recipes: recipeViewModel.toCookRecipes
-                        )
-                    }
-
-                    if !recipeViewModel.cookedRecipes.isEmpty {
-                        RecipeSection(
-                            title: RecipeStatus.cooked.displayTitle,
-                            recipes: recipeViewModel.cookedRecipes
-                        )
-                    }
-                }
-                .padding(.vertical)
+        ScrollView {
+            HStack {
+                Spacer()
+                NavigationLink(destination: CreateRecipeView()) {
+                    Image(systemName: "plus")
+                        .imageScale(.large)
+                        .foregroundStyle(.tint)
+                    Text("Add New")
+                }.padding(.trailing)
             }
-            .navigationTitle("Recipes")
+
+            VStack(spacing: 24) {
+                if !recipeViewModel.cookingRecipes.isEmpty {
+                    RecipeSection(
+                        title: RecipeStatus.cooking.displayTitle,
+                        recipes: recipeViewModel.cookingRecipes
+                    )
+                }
+
+                if !recipeViewModel.toCookRecipes.isEmpty {
+                    RecipeSection(
+                        title: RecipeStatus.toCook.displayTitle,
+                        recipes: recipeViewModel.toCookRecipes
+                    )
+                }
+
+                if !recipeViewModel.cookedRecipes.isEmpty {
+                    RecipeSection(
+                        title: RecipeStatus.cooked.displayTitle,
+                        recipes: recipeViewModel.cookedRecipes
+                    )
+                }
+            }
+            .padding(.vertical)
         }
+        .navigationTitle("Recipes")
 
     }
 }
